+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix automated Drupal 10 compatibility fixes - Issue #3329302 by Project Update Bot
 - fix library testing path that may be inconsistent between Github Actions & GitlabCI
 - fix call to deprecated method withConsecutive() on PHPUnit
+- fix usage of deprecated HttpKernelInterface::MASTER_REQUEST use ::MAIN_REQUEST instead
 
 ## [1.0.0] - 2022-12-16
 ### Fixed